+++ src/Angular2Apollo.ts
@@ -1,12 +1,13 @@
 import { OpaqueToken, Injectable, Inject } from '@angular/core';
 import { rxify } from 'apollo-client-rxjs';
 import { ApolloQueryResult } from 'apollo-client';
+import { Observable } from 'rxjs/Observable';
 
 import { ApolloQueryObservable } from './ApolloQueryObservable';
 
 import ApolloClient from 'apollo-client';
 
-import 'rxjs/add/operator/switchMap';
+import 'rxjs/add/observable/from';
 
 export const angularApolloClient = new OpaqueToken('AngularApolloClient');
 export const defaultApolloClient = (client: ApolloClient): any => {
@@ -22,15 +23,23 @@ export class Angular2Apollo {
     @Inject(angularApolloClient) private client: any
   ) {}
 
-  public watchQuery(options): ApolloQueryObservable<ApolloQueryResult> {
+  public watchQuery(options: any): ApolloQueryObservable<ApolloQueryResult> {
     return new ApolloQueryObservable(rxify(this.client.watchQuery)(options));
   }
 
-  public query(options) {
+  public query(options: any) {
     return this.client.query(options);
   }
 
-  public mutate(options) {
+  public mutate(options: any) {
     return this.client.mutate(options);
+  }
+
+  public subscribe(options: any): Observable<any> {
+    if (typeof this.client.subscribe === 'undefined') {
+      throw new Error(`Your version of ApolloClient doesn't support subscriptions`);
+    }
+
+    return Observable.from(this.client.subscribe(options));
   }
 }